Remove HEMCO_Config.template file and instead prompt user to provide path to a file
Instead of maintaining a separate HEMCO_Config.rc file in this repository, the
createRunDir.sh script prompts the user to provide a path to a HEMCO_Config.rc
file from a run directory of HEMCO_Config.template file from the GEOS-Chem
unit tester (or soon-to-be located in the code repository).  The script has
been updated to replace tokens in HEMCO_sa_Config.rc and HEMCO_Config.rc as
needed.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>