Add fix to hcoio_read_std_mod.F to exit search only if a field is to be interpolated and isn't discontinuous
For the implementation of Yuan-processed MODIS LAI in GEOS-Chem 12.7.0,
code was added to hcoi_read_std_mod.F90 to exit the search for the right
timestamp. The search was supposed to be used only for non-interpolated
data, but it turns out that code is needed to properly read in met fields
for the HEMCO standalone. As a fix, we only exit the search if the dataset
is to be interpolated and is not discontinuous (as is the case for Yuan-
processed MODIS LAI).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>