Add back in Makefile to make identical to GEOS-Chem 12.8.1 HEMCO directory
This Makefile is not used in the HEMCO repository. HEMCO should only
be built with CMake. However, the current version of GEOS-Chem still
contains a copy of HEMCO as a subdirectory. To facilitate development of
GEOS-Chem and HEMCO repositories at the same time we are making the HEMCO
repository and the HEMCO subdirectory in GEOS-Chem identical. This will
continue until GEOS-Chem 13.0.0 when we switch to using the HEMCO
repository exclusively and retire GNU make.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>