Rename Shared modules to be unique from GEOS-Chem
This avoid name conflicts when building GEOS-Chem and HEMCO
together and moves towards making HEMCO aesthetically independnet
from GEOS-Chem by consistently using module names HCO_xx.

The only Shared file not touched in this update is precision_mod.F90.
That file should ultimately be removed and all files in HEMCO use
HEMCO precision. This will be a future update.

For now these files are still in the same directory structure as
prior to this update.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>