Do not populuate HEMCO data arrays for dry-run simulations
In HEMCO/src/Core/hcoio_read_std_mod.F90, we have removed calls to
FileData_ArrInit that populate arrays with dummy values when doing
a GEOS-Chem or HEMCO dry-run simulation.  This causes an excessive
amount of memory to be allocated.

With this fix, even a geosfp_2x25_standard benchmark now only requires
5.3 GB of memory (instead of 32GB, as was the case previously).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>