Removing HCO_Enter from subroutine missing HCO_Leave
HCO_Enter is called within gethcodiagn but the analogous
HCO_Leave is not called. This results in the "maxnest too
low" error since the maximum number of nested subroutines
tracked for tracedback is exceeded. Rather than add in
HCO_Leave, HCO_Enter is removed. These subroutines are
meant for "big" calling subroutines only.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>