Dry-run bug fix: exit if the current year is outside the specified range
We have added a more robust error trap to exit the GEOS-Chem and/or
HEMCO dry-run when the current year lies outside the specified year
range of a given field in the HEMCO_Config.rc file.

For example, this situation can happen if you have met fields specified
with years 1980-2019 in HEMCO_Config.rc, but then try to do a dry-run
for 2020.

The code now prints an error message to the log file and exits gracefully.

This commit corresponds to geoschem/geos-chem afce43f.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>