Remove HCO_UFLAG_ONCE from EFY time cycle option in HEMCO
The EFY flag is used to ensure the exact dates are used for meteorology
fields and boundary conditions. For some reason, the E time cycle options
all include Dta%UpdtFlag = HCO_UFLAG_ONCE, which causes the fields to only
be read once at the start of the simulation. It's not clear why we'd want
to only update once. For EFY this option has been removed, but it remains
for E, EF, etc. to avoid breaking things. We can investigate and clean up
these flags at a later date.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>