Make sure NO and NO2 are positive when computing FNO_NOx
In routine Evolve_Plume (in HEMCO module src/Extensions/paranox_mod.F90),
we need to make sure that both NO and NO2 are nonzero before computing
FNO_NOx.  This involves changing an .or. to an .and. in the IF block.
Otherwise, set FNO_NOx = 1.0.

This commit resolves the issue brought up in:
https://github.com/geoschem/geos-chem/issues/380

This update corresponds to commit ff0a547 in geoschem/geos-chem.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>