Now check standard_name and long_name attributes for sigma levels
In module src/Shared/NcdfUtil/hco_ncdf_mod.F90, we have made the
following updates to the algorithm that determines if the data is
placed on sigma levels:

(1) In routine NC_Get_Sigma_Levels_C, we now return an array of sigma
    values if either the "standard_name" or "long_name" attribute of
    the level variable is "atmosphere_hybrid_sigma_pressure_coordinate".

(2) In routine NC_IsSigmaLevel, we return TRUE if the "standard_name"
    or "long_name" attribute of the level variable contains the
    text "atmosphere_hybrid_sigma_pressure_coordinate".

These updates should now cause HEMCO to vertically interpret emissions
and to not place all the emissions into the surface layer.

Also made some cosmetic changes to mirror the code in the NcdfUtil
folder of the GEOS-Chem repository.

The corresponding commit in the GEOS-Chem repository is:
https://github.com/geoschem/geos-chem/commit/2e46145c7abfe5c66c8d0a1c9262e472610294be

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>