Restore EmisNO_Fert diagnostic in hcox_soilnox_mod.F90
In GEOS-Chem 12.7.0, the capability to save out fertilizer NOx emissions
separately from soil NOx was inadvertently removed. In that version, a
majority of the code for bpch diagnostic was removed and this included
code for creating manual HEMCO diagnostics in hcoi_gc_diagn_mod.F90. In
that module, code for creating the EmisNO_Fert diagnostic was also removed
but was still needed. With this update, the EmisNO_Fert diagnostic is now
created and updated automatically (via the Inst%FertNO_Diag pointer) in
hcox_soil_nox_mod.F90.

The following line is needed in HEMCO_Diagn.rc to save out the fertilizer
NOx emissions:

  EmisNO_Soil        NO     104    -1  -1   2   kg/m2/s  NO_emission_flux_from_soil_including_fertilizer
  EmisNO_Fert        -1     104    -1  -1   2   kg/m2/s  NO_emission_flux_from_fertilizer_only

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>