Fix bug in xyL option in HEMCO standalone if box height not available
Using the xyL option for 2D emissions should apply the total emissions
up to level L using a uniform distribution, thereby preserving the total.
A bug in the subroutine to calculate the dilution factor used to scale
the emissions instead caused duplication of total at each level when box
height was unavailable. This update corrects this issue by ensuring the
box height value is not associated when box height is not available.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>