Ensure box height is initialized for calculation if pressures available
This update builds on top of a recent commit that nullified the
box height pointer if box height was not available externally. For the
case of pressure being available, however, if box height is not available
then it will be calculated from the pressure. The previous update
introduced a seg fault for that case. This commit fixes that.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>