Remove Interfaces/GCHPctm directory since unnecessary
With the decision to not use HEMCO as a proper ESMF gridded component in
GCHP, the placeholder for storing the grid comp fortran file is no longer
needed. Rather than link to library HCOI_GCHP, GCHPctm and GEOS-Chem now
link to HCOI_MAPL_ESMF, and the folder with HCOI_GCHP defined is removed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>