Remove restriction that MATH expressions must contain a time value
For the unit conversion from kgC to kg species, it would be ideal to
provide the expression used to compute the scaling rather than a hard-
coded value. The MATH capability in HEMCO can be of use here, but we need
to remove the restriction that MATH expressions must include a time
value. This should make it more clear what values are being used to
generate scale factors. For example, in HEMCO_Config.rc we can now use:

   # Units carbon to species conversions
   # Factor = # carbon atoms * MW carbon) / MW species
   40 CtoACET MATH:58.09/(3.0*12.01)   - - - xy unitless 1

which will return in HEMCO.log (with Verbose on):

   Read from config file: CtoACET
   Evaluated function: 58.09/(3.0*12.01) -->    1.61226755481543

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>