Add fixes to MEGAN extension following removal of carbon-based units
The first round of internal benchmarking revealed larger than expected
differences for ACET, SOAP, and SOAS emissions from MEGAN. To resolve the
differences in ACET, we need to account for the unit conversion from
MONO to ACET and MBOX to ACET since we are no longer working in units of
carbon. Including this conversion brings values closer to pre-carbon unit
removal but there are still differences because the old code assumed a
molecular weight of 120 g/mol for MTPA, 36 g/mol for ACET, and  60 g/mol
for MBOX (i.e. MW of carbon times number of carbon). The updated code uses
the more precise values of 136.26 g/mol for MTPA, 58.09 g/mol for ACET, and
86.13 for MBOX.

In order to reproduce biogenic SOAP and SOAS versions from before the carbon
unit removal, we need to convert emissions of isoprene, monoterpenes, and
other terpenes to kgC before applying the scale factors to convert to SOAP
and SOAS. New scale factors may need to be computed to properly account for
emissions in units species but this works for now.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>