HCO_CalcVertGrid now looks for PS and TMPU fields as well
If arguments PSFC and TK are null in HCO_CalcVertGrid, it will attempt
to read those fields from the HEMCO state.  However, PS and TMPU are
also often used in place of PSFC and TK.  Therefore, if PSFC is not
found, it will look for PS and if TK is not found, it will look for
TMPU.

This will make sure that the HEMCO standalone gets pressure and
temperature fields that are necessary to compute other vertical
quantities.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>