Remove DYNAMIC & COLLAPSE from hco_calc_mod.F90 and hco_geotools_mod.F90
We have reverted to OMP static loop scheduling, as using DYNAMIC, 1
actually incurs more overhead than not using it.

We have also restored EXIT statements in routine GET_CURRENT_EMISSIONS.
It is OK to use them because these will exit from the "I" loop, which
is not parallelized. Only the "J" loop is parallelized.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>