Add new time cycle flag "EFK" for reading met fields w/ FlexGrid
In the prior commit, the time cycle flag "EFY" was updated to
"Exact Year, Force error if not found, Use simulation Year, read once".
This was intended for reading GEOS-Chem "Classic" restart files.

An unintended side-effect of this modification was that it caused
met fields (which also were tagged with the EFY time cycle flag) to
only be read once instead of continuously being read throughout the
simulation.

We have introduced a new flag "EFYK", which denotes "Exact year,
Force error if not found, use simulation year, keep on reading).
This should be used for met field files that are read with FlexGrid.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>