Bug fix: Use DO I=1,N for interpolated data in routine Check_AvailYMDhm
A previous fix in HEMCO 3.0.0-rc.3 changed DO I = 1, N (which would
have caused an out-of-bounds error) to DO I = 1, N-1 in routine
Check_AvailYMDhm (in hcoio_util_mod.F90).  However, for
interpolated data (such as Yuan_XLAI), this causes the check for
the closest time point to fail.

We have put in a check that sets the upper loop limit to N for
interpolated data (if HCO_CFLAG_INTER = True), and to N-1 otherwise.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>