Add capability to copy emissions vertically by specifying dimensions xyL* in HEMCO_Config.rc
In order to vertically allocate emissions via scale factors, we first need
to copy emissions vertically. Scale factors can then be applied to determine
how much emissions to apply to each level. For example, in HEMCO_Config.rc
we define a file vert_alloc.nc that defines the ratio of emissions to apply
to each level for CEDS energy, industry, and ship emissions.

  #==============================================================================
  # --- CEDS vertical partitioning ---
  #==============================================================================
  (((CEDS
  315 ENERGY_LEVS   vert_alloc.nc g_energy   2017/1/1/0 C xyz 1 1
  316 INDUSTRY_LEVS vert_alloc.nc g_industry 2017/1/1/0 C xyz 1 1
  317 SHIP_LEVS     vert_alloc.nc cmv_c3     2017/1/1/0 C xyz 1 1
  )))CEDS

These scale factors are then applied to the CEDS_*_ENE, CEDS_*_IND, and
CEDS_*_SHIP fields. These fields are 2D in the CEDS files, but we now
can specify dimension xyL* instead of xy to tell HEMCO to copy the field
into each emissions level:

  0 CEDS_CO_ENE $ROOT/CEDS/v2020-08/$YYYY/CO-em-total-anthro_CEDS_$YYYY.nc  CO_ene  1970-2017/1-12/1/0 C xyL* kg/m2/s CO 26/37/35/315 1  5
  0 CEDS_CO_IND $ROOT/CEDS/v2020-08/$YYYY/CO-em-total-anthro_CEDS_$YYYY.nc  CO_ind  1970-2017/1-12/1/0 C xyL* kg/m2/s CO 26/316       1  5
  0 CEDS_CO_SHP $ROOT/CEDS/v2020-08/$YYYY/CO-em-total-anthro_CEDS_$YYYY.nc  CO_shp  1970-2017/1-12/1/0 C xyL* kg/m2/s CO 26/317       10 5

In hco_config_mod.F90, a check for xyL* is done and if found, it sets
the first emission level to 1, the second emission level to 0 (temporarily)
with units = HCO_EMISL_TOP, and the emission level mode to 2.

In hco_calc_mod.F90, routine HCO_GetIDx has been updated so that if level
units = HCO_EMISL_TOP then the level index is set to HcoState%NZ (or the
top emissions level). Then in routine Get_Current_Emissions, when
determining how to distribute emissions vertically, if EmisLMode = 2 then
the emissions will be copied to each level. Otherwise, if 2D emissions
are being converted to 3D then a dilution factor will be computed based
on grid box height. Finally, in that code, the vertical scale factors are
applied to properly distribute the emissions for each level.

NOTE: To properly save out these emissions, the dimensions in HEMCO_Diagn.rc
should be changed from dimension = 2 to 3. For example:

  EmisCO_Anthro      CO     0    1   -1   3   kg/m2/s  CO_emission_flux_from_anthropogenic
  EmisCO_Ship        CO     0    10  -1   3   kg/m2/s  CO_emission_flux_from_ships
  InvCEDS_CO         CO     0    1   5    3   kg/m2/s  CO_emission_flux_from_CEDS_inventory

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>