Update createRunDir.sh to replace tokens in HEMCO_Config.rc recently introduced by GEOS-Chem updates
The createRunDir.sh script in this repository has been updated for
consistency with the analagous script in geoschem/geos-chem. This includes
replacing tokens primarily introduced by the GCAP 2.0 update in
GEOS-Chem 13.1.0. As part of this, the GCAP 2.0 is now also an option
for the meteorology fields in the HEMCO standalone.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>