Send all HCO_ERROR messages to stderr and write from all threads
This update is a quick fix to inconsistency writing HEMCO error messages
to log. There were several issues:

1. HCO_ERROR was sometimes called to pass error messages to HEMCO.log and
   other times called to pass error messages to stdout. This leads to
   confusion about where to look for error messages.

2. If the error message was passed to HEMCO.log then it was only written
   by the root thread. Because HCO_ERROR is called right before early
   program termination the root thread would sometimes not get to write
   the message before early exit. This was especially true for MPI
   applications with many cores.

3. Passing the error message to HEMCO.log requires that the HEMCO.log is
   open to write to. But the HEMCO.log file is only opened by the root
   thread and other processes can get ahead of that if using many cores.

The subroutine that HCO_ERROR invokes if passing messages to HEMCO.log
is still present in the code but will be reassessed in the future. For now,
writing messages when there is early termination to std out and by all
threads is the most straightforward and consistent way to alert users
to the source of the error.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>