Add climatology input to volcano extension
In HEMCO_Config.rc we now specify 'Volcano_Climatology' which is the path
to a file containing the degassing volcano climatology data.

   117     Volcano                : on    SO2
       --> Volcano_Source         :       AeroCom
       --> Volcano_Table          :       $ROOT/VOLCANO/v2021-09/$YYYY/$MM/so2_volcanic_emissions_Carns.$YYYY$MM$DD.rc
       --> Volcano_Climatology    :       $ROOT/VOLCANO/v2021-09/so2_volcanic_emissions_CARN_v202005.degassing_only.rc

The volcano extension will first look for a volcano file for the current
simulation date. If that file is not found, it will then attempt to look
for the climatology file. If neither file is found, the simulation will
crash with an error.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>