Modify volcano extension so that dry-run option also looks for climatology file
The code in hcox_volcano_mod.F90 has been modified so that a dry-run simulation
will also proceed to look for the volcano climatology file is a file is not
found for the current simulation date. This will ensure that both the volcano
emissions file for the simulation date and the volcano climatology file are
downloaded by the dry-run option since it will always be useful to have the
climatology file locally.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>