Fix indexing of species in hcox_seasalt_mod.F90 when using marinePOA option
When using the marinePOA option in the sea salt extension, an out-of-bounds
error occurs in SpcNamesSS. In GEOS-Chem 12.9.0, the Br2 option was retired
from this extension and the Br2 species was removed from the SpcNamesSS
array. However, the WRITE statements using that array in HCOX_SeaSalt_Init
were never updated to use the new species index values. This is now fixed.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>