Bug fix: Restore COL=-1 in calls to Diagn_Update in hco_calc_mod.F90
Passing COL=-1 to Diagn_Update in hco_calc_mod.F90 is needed to update
all HEMCO diagnostics, not just the default diagnostics. This is essential
to update manual diagnostics, such as those defined in GEOS-Chem's
hcoi_gc_diagn_mod.F90 for the CH4, Hg, and TOMAS simulations. Without
this fix, arrays containing all zeroes are passed when manual diagnostics
are obtained via HCO_GC_GetDiagn.

This fix reverts changes made in commit https://github.com/geoschem/HEMCO/commit/057363277fc0820ee7d4828ccbd23c4308b9992b
for the GCHP adjoint. It is not clear if this has an impact on the GCHP adjoint,
but if so then changing the value of COL should be done within an IF block
so that it only applies to adjoint simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>