Fixed syntax error in GET_ORO routine of hcox_tomas_dustdead_mod.F
In a previous modification to HEMCO, the FRLANDICE field of ExtState
is used in routine GET_ORO.  However, this modification caused a code
line to go beyond 72 characters, which is the limit for fixed-format.

We have deleted some whitespace in this line of code so that all code
does not go beyond column 72.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>