GCHP bug fix: revert to all zeros in criteria for restart field not filled
This commit moves a recent change introduced for GEOS to MODEL_GEOS blocks
and restores legacy handling used in other ESMF models. The GEOS code uses all
negatives to determine if a restart field was bootstrapped (not present
in restart file so auto-filled with values in internal state). The legacy
code uses all zeros. The legacy approach is problematic because
only part of the globe is assessed during MPI runs, and it is possible
for a restart variable to be all zeros in part of the globe and non-zero
elsewhere.

For now the legacy approach is restored for a quick fix for GCHP. The
approach taken in non-GEOS runs will be made more robust in the next
GCHP version.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>