Bug fix: Initialize Dta%Discontinuous = .FALSE. in hco_config_mod.F90
This commit fixes the issue reported in geoschem/HEMCO #132.  We had
observed a difference in output in runs using different numbers of
computational cores.  We had thought this was a parallelization issue
but in fact, the interpolation of leaf area indices was being done
incorrectly.

The root cause of the issue was that the Dta%Discontinuous field was
being used before it was initialized.  In some instances, this was
inadvertently assigned a TRUE value when it should have been FALSE.
Adding a Dta%Discontinuous = .FALSE. statement above the IF block
where it is used fixes the problem.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>