Abstract handling of time cycle flags & related code out of Config_ReadCont
In routine Config_ReadCont (in module src/Core/hco_config_mod.F90),
there is duplicated identical code in both an IF and ELSE block, which:

(a) Handling time cycle flag information
(b) Determining the dimension of the data array vector as 2D or 3D
(c) Setting the level scale factor indices
(d) Checking if a mask is assigned to a scale factor
(e) Extracting grid box edges

This has proven to be problematic, and a source of error, as often
one instance of the duplicated code is modified but the other is not.

To correct this issue, we have abstracted the above-mentioned code
into a separate routine called UpdateDtaProperties, which is called
from both locations.

This modification gives identical results except for numerical that seems
to be caused in the PARANOx extension.  We shall try to resolve this
in a subsequent commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>