Restore logic for copying HEMCO_Config.rc files to createRunDir.sh
The previous commit changed the logic for copying HEMCO_Config.rc files
and removed the wildcard character * from the copy command. The wildcard
is needed to copy the HEMCO_Config.rc containing the met field entries
(e.g. HEMCO_Config.rc.gmao_metfields) if it exists. That line is now
restored here. In addition, the IF statement is also reverted so it
checks for the existence of the HEMCO_Diagn.rc file as originally intended.
The check for the existence of HEMCO_Config.rc is already done right after
the user specifies the path to that file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>