Increase the string length for reading lines from HEMCO standalone grid file
In order to run the HEMCO standalone with the 0.25x0.3125 global grid
(defined in run/HEMCO_sa_Grid.025x03125.rc), the line length for reading
YEDGE and YMID must be increased. Previously, it was LEN=4095 when the
length of YEDGE is 5341 and YMID is 4616. To prevent errors when reading
the HEMCO standalone grid file and defining the grid in routine Set_Grid,
the length of DUM has been increased to 5500.

This addresses https://github.com/geoschem/HEMCO/issues/206.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>