Fix bug introduced by prior commit of blocking prints to root thread only
This update fixes a problem where the simulation will crash in the
volcano extension init subroutine on non-root threads if the volcano
extension is turned off, such as in the transport tracer simulation. The
fix moves the return if extension number is zero to directly after the
call to retrieve extension number. This is the same handling as all other
HEMCO extension initialization routines.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>