Turn off emission extensions when EMISSIONS logical is set to false
The EMISSIONS logical switch in HEMCO_Config.rc currently only turns off
base emissions and not the emission extensions. The emission extension data
are bracketed by the logical so those data are not read in. However, because
the extensions themselves are still on, they get called in HEMCO and result
in the model crashing because the input data can't be found.

This is now fixed by adding a check in subroutine ExtSwitch2Buffer (in
hco_config_mod.F90) to see if EMISSIONS is set and using that to define a
local logical DoEmis. In that same routine, if DoEmis is false then all
emission extensions will be defined with Enabled = .FALSE. to avoid calling
them during the run stage of HEMCO.

Addresses https://github.com/geoschem/HEMCO/issues/249.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>