Add field outLUN to HcoConfig object to store output LUN
Default is 6 if not passed to ConfigInit. In GEOS-Chem this corresponds
to standard output. This enables specifying destination other than stdout
or HEMCO.log for HEMCO prints in models. For example, in CESM this allows
printing to atm.log rather than cesm.log (stdout) by using
HcoConfig%outLun in write statements.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>