Increase string length for reading lines from HEMCO standalone grid file
The HEMCO standalone currently crashes when running global 0.125x0.15625 (aka 12km) resolution simulations,
because the YEDGE and YMID lines in https://github.com/geoschem/HEMCO/blob/main/run/HEMCO_sa_Grid.0125x015625.rc
are too long.

In PR #207 we increased the maximum string length to 5500 to allow for global 0.25x0.3125 HEMCO standalone simulations.
However for 0.125x0.15625 global grids, the length of the YEDGE and YMID string in HEMCO_sa_Grid.rc are 12,103 and 10,657
characters, respectively. To get around this we can increase the length of string DUM from 5500 to 15000.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>